gasgauge patching code
So, it turns out that not only do we need to set capacity, we need
to set total energy, /and/ we have to check every time on boot.
Because the gas gauge loses its nuts when the battery is
disconnected, and goes back to a default capacity which is
inaccurate.

There's no way for us to know if we lost our battery, so every
time on boot we have to unseal the gauge and check its status.
Unfortunately, if you unseal multiple times rapidly, the system
locks you out.

So we have a retry sequence with UI prompt to indicate when this
is happening, for example, if someone resets the system twice
in rapid succession.