heavy debugging of gasgauge code
Added routines to print a lot more data about what's going through
the gasgauge's mind.

The problem appears to be that the gasgauge is "giving up"
on trying to track a new programming if it can't get all the
measurements it wants, e.g. OCV and "resim", whatever that is,
and going back to the defaults.

Reprogramming the defaults and hitting a soft reset forces
OCV and resim, which looks like it zeroes out all the tracking
information.

Trying a variant where to exit programming, we simply seal
and exit without doing a reset, OCV, or resim. Let's see if
that fixes it....