much more gasgauge debugging.
-added command "ggdump" which dumps most of the state of the
gasgauge so we can compare a broken gauge to a working one

-added command "ggfix" which fixes the params that seem to be
going wrong.

Basically, it looks like either Qmax (in subclass 82) documentation
is wrong, or the gas gauges aren't dealing well with it, due to
some sort of math overflow error. It's supposed to hover around
16384, but it seems to find it's way around +/- 0. It's totally
reasonable that Qmax might be encoded not in the ratiometric
form suggested in the doc and instead as a +/- mAh offset from
base capacity. It would even make a whole lot more sense to do
it that way.

It also looks like, for some reason, on one board the CC offset term
was thrown way out. So if we find that we're constantly resetting
ourselves to the wrong capacity values, restore both Qmax and
cal values.

This seems to have resurrected at least one misbehaving gas gauge.
Only time and more testing will tell for sure.