fix critical issue with gasgauge: timeout after selecting data block
There's a bug in the gas gauge. After setting a data block to
read, if you don't wait about 3ms, the first byte of the read
out block is corrupt. This will affect the gasgauge update routine
because it reads the whole block, mods, and re-writes it.

Also tweak more parameters & values; set TERM current to be higher
(because we cut-off at 0.05C, 200mA, the taper isn't being caught)
and also adjust the discharge/charge/quit thresholds to reflect
the application.